dsl(groovy): use java to implement the route buiolder loader instead of groovy
