+++ dsl/camel-groovy-dsl/camel-groovy-dsl-test/src/test/resources/log4j2-test.properties
