+++ dsl/camel-groovy-dsl/camel-groovy-dsl-test/src/test/resources/routes/issues/github-529.groovy
