+++ dsl/camel-groovy-dsl/camel-groovy-dsl-test/src/test/resources/routes/routes-with-endpoint-dsl.groovy
