+++ parent/pom.xml
@@ -2726,6 +2726,11 @@
 				<artifactId>camel-groovy-dsl</artifactId>
 				<version>${project.version}</version>
 			</dependency>
+			<dependency>
+				<groupId>org.apache.camel</groupId>
+				<artifactId>camel-groovy-dsl-common</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 
             <!-- camel catalog -->
             <dependency>
